requirements.txt
discord.py==1.7.3
Pillow==10.3.0
python-dotenv==0.19.0